refactor(cio): expose evaluate_scout_report service for ROB-196 e2e
Extract the G1~G6 gate sweep from scripts/cio_quality_gate.py into
app.services.cio_quality_gate_service so ROB-196 e2e fixtures can
import and assert on the same engine the CIO CLI runs.

Public API (pins the ROB-196 contract):

    report = evaluate_scout_report(
        markdown=report_md,
        cash_balance=1_670_000,                            # raw KRW, optional
        tool_failures=["screen_stocks: schema mismatch"],  # optional
    )
    report.overall_status  # "PASS" | "PARTIAL" | "FAIL"
    report.violations      # list[GateViolation(gate_id, severity, detail)]
    report.reopen_comment  # str | None  (set on FAIL)

- app/services/cio_quality_gate_service.py: new module, owns all parsing,
  gate logic, reopen builder, and the render_report() helper. Keeps the
  legacy GateResult dataclass for the CLI renderer and adds GateViolation
  + QualityGateReport for the ROB-196 assertion surface.
- scripts/cio_quality_gate.py: now a thin CLI wrapper (~160 LOC, down
  from ~690). Accepts --cash in raw KRW to match the service API, plus
  --tool-failure to inject extra G3 signals.
- tests/test_cio_quality_gate.py: drop the importlib module-loader hack,
  import directly from app.services.cio_quality_gate_service. Keep the
  six ROB-158 smoke tests and add three evaluate_scout_report() tests
  covering overall_status, cash_balance override, and merged
  tool_failures.

Note: ROB-196 QA branch scaffolds tests/test_cio_quality_gate.py with
five additional scout_reports fixtures (g1/g3/g4/g6/all_gates_pass). The
file path overlaps this commit; merge will need to concatenate assertions.
